shopping cart

+++ ShoppingCart/ClsShoppingCart.cs
@@ -2,6 +2,7 @@
 using Microsoft.AspNetCore.Components.Server.ProtectedBrowserStorage;
 using Services.Interfaces;
 using System;
+using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
@@ -13,74 +14,48 @@ namespace ShoppingCart
         private CartItem _CartItem { get; set; }
         private readonly IProduct ProductServices;
         private readonly ProtectedLocalStorage LocalStorage;
-
- 
-
-        public ClsShoppingCart()
-        {
-        }
+        public ClsShoppingCart() { }
         public ClsShoppingCart(IProduct ProductServices, ProtectedLocalStorage LocalStorage)
         {
             this.ProductServices = ProductServices;
             this.LocalStorage = LocalStorage;
         }
-
         public ClsShoppingCart(Cart cart, CartItem cartItem)
         {
             _Cart = cart;
             _CartItem = cartItem;
-         }
+        }
         public async Task<Cart> InitializeShoppingCart()
         {
-
             var result = await LocalStorage.GetAsync<Cart>("Cart");
-
             return _Cart = result.Success ? result.Value : null;
-
-
-
-
         }
-
-
         public async Task<Cart> AddCart(int id)
         {
-            //si null je fait une instance
-
             if (_Cart == null)
             {
                 _Cart = new Cart();
             }
-
-            // je recupere le produit que je dois ajouter a Cart
             var products = await ProductServices.GetAllProductsIncludePicturesAsync();
             var productById = products.ToList().Where(a => a.IdProduct == id).FirstOrDefault();
-            //je crée CartItem
             _CartItem = _Cart.ListProducts.Where(a => a.ProductId == productById.IdProduct).FirstOrDefault();
-            //verif
-            //ajout cartItem a Cart
             if (_CartItem != null)
             {
                 foreach (var session in _Cart.ListProducts.ToList())
                 {
                     if (session.ProductId == productById.IdProduct && session.ProductPrice != productById.SalesPrice)
                     {
-
                         session.ProductPrice = productById.SalesPrice;
-
                     }
                 }
                 if (productById.Quantity > _CartItem.ProductQuantity)
                 {
                     _CartItem.ProductQuantity++;
-
                 }
                 _CartItem.TotalItem = _CartItem.ProductPrice * _CartItem.ProductQuantity;
-
             }
             else
             {
-
                 _Cart.ListProducts.Add(
                 new CartItem()
                 {
@@ -92,18 +67,13 @@ namespace ShoppingCart
                     ProductQuantity = 1,
                     TotalItem = productById.SalesPrice,
                     ProductUrlImage = productById.ImageUrl,
-
                 });
             }
-
-            //calcul total...
             _Cart.Total = _Cart.ListProducts.Sum(a => a.TotalItem);
             _Cart.TotalAllItems = _Cart.ListProducts.Sum(t => t.TotalItem);
             await LocalStorage.SetAsync("Cart", _Cart);
-             return _Cart;
+            return _Cart;
         }
-
-
         public async Task ClearCart()
         {
             if (_Cart != null)
@@ -114,12 +84,10 @@ namespace ShoppingCart
                 }
                 _Cart.TotalAllItems = _Cart.ListProducts.Sum(a => a.TotalItem);
                 _Cart.Total = _Cart.ListProducts.Sum(a => a.TotalItem);
-                await LocalStorage.SetAsync("Cart", _Cart);
                 await LocalStorage.DeleteAsync("Cart");
-
+                await LocalStorage.SetAsync("Cart", _Cart);
             }
         }
-
         public async Task DeleteCartItem(int id)
         {
             _CartItem = _Cart.ListProducts.FirstOrDefault(a => a.ProductId == id);
@@ -127,52 +95,40 @@ namespace ShoppingCart
             _Cart.TotalAllItems = _Cart.ListProducts.Sum(a => a.TotalItem);
             _Cart.Total = _Cart.ListProducts.Sum(a => a.TotalItem);
             await LocalStorage.SetAsync("Cart", _Cart);
-            if (_Cart.ListProducts.Count().Equals(0))
+            if (_Cart.ListProducts.Count() <= 0)
             {
                 await LocalStorage.DeleteAsync("Cart");
-
+                await LocalStorage.SetAsync("Cart", _Cart);
             }
-
-
         }
-
-
         public async Task DeleteItem(int id)
         {
             if (_Cart != null && _Cart.ListProducts.Count > 0)
             {
                 _CartItem = _Cart.ListProducts.FirstOrDefault(a => a.ProductId == id);
-
             }
-            if (_CartItem != null)
+            if (_CartItem != null && _CartItem.ProductQuantity > 0)
             {
-                if (_CartItem.ProductQuantity > 0)
-                {
-                    _CartItem.ProductQuantity = _CartItem.ProductQuantity - 1;
-
-                    _CartItem.TotalItem = _CartItem.TotalItem - _CartItem.ProductPrice;
-
-                    //calcul total...
-                    _Cart.Total = _Cart.ListProducts.Sum(a => a.TotalItem);
-                    _Cart.TotalAllItems = _Cart.ListProducts.Sum(t => t.TotalItem);
-                }
-
-                if (_CartItem.ProductQuantity == 0)
+                _CartItem.ProductQuantity = _CartItem.ProductQuantity - 1;
+                _CartItem.TotalItem = _CartItem.TotalItem - _CartItem.ProductPrice;
+                _Cart.Total = _Cart.ListProducts.Sum(a => a.TotalItem);
+                _Cart.TotalAllItems = _Cart.ListProducts.Sum(t => t.TotalItem);
+                await LocalStorage.SetAsync("Cart", _Cart);
+                if (_CartItem.ProductQuantity <= 0)
                 {
                     await DeleteCartItem(_CartItem.ProductId);
                 }
-                await LocalStorage.SetAsync("Cart", _Cart);
-
             }
-
-
         }
-
-        public Cart getAll()
+        public async Task CheckDisponibility(List<Domains.Product> allProductsEmpty)
         {
-
-
-            return _Cart;
+            if (allProductsEmpty != null)
+            {
+                foreach (var product in allProductsEmpty.Where(a => a.Quantity <= 0).ToList())
+                {
+                    await DeleteItem(product.IdProduct);
+                }
+            }
         }
     }
 }
